requirements.txt
# Airflow DB
psycopg2-binary

# Spark
pyspark
ipykernel
pandas
seaborn
matplotlib
numpy
scikit-learn
fastapi 
sqlalchemy 
uvicorn
pydantic
joblib
python-dotenv 
pytest
python_jose
passlib[argon2]
requests
apache-airflow
pyarrow
pyarrow
pydantic[email]
pwdlib
